backend/apps/site_config/api/serializers.py
from rest_framework import serializers

from apps.site_config.models import SiteSetting


class SiteSettingSerializer(serializers.ModelSerializer):
    class Meta:
        model = SiteSetting
        fields = "__all__"
